fix: render terrain, unblock mission load, and steady the cursor
The ground was black because the SDL GPU adapter's GetViewMatrix returned the
direct-transform matrix while a direct transform was active. CScene::Draw draws
terrain under one deliberately, and CTerrain::MovePatches transforms the terrain
origin through viewport * projection * view to find where the map begins on
screen. With the direct matrix in that product it collapses to the identity, so
the origin came back as the world origin, patches landed thousands of pixels
below the window and AddVertices clipped away every tile. The legacy engine
returns the stored camera view here, and matching it takes the tile meshes from
zero vertices to 1708 noise tiles, 1536 plain tiles and 976 crosses.

Two failures had to be fixed before a mission would load at all:

lmem.c enabled Lua's testing allocator unconditionally. Game, libUI and
libAILogic each link a private copy of it while luaM_realloc is exported, so
blocks allocated through one module's counters were freed through another's.
Those counters are unsigned and wrapped -- measured at 2^64-695 bytes and
2^64-16 blocks -- after which every Lua allocation was refused and lua_close
called exit(1) mid load. Its asserts are compiled out in release, so it died
silently.

SIntPairHash computed v.x * RAND_MAX, which is 32767 on Windows but INT_MAX
here, overflowing a signed int and tripping the UBSan trap. It is unsigned now,
which leaves the Windows result unchanged.

Vertex layouts now come from the FVF instead of one hardcoded 32-byte stride
with the colour at 12 or 16, which described only SGFXLVertex and SGFXTLVertex
correctly. Terrain tiles are 36 bytes, mesh vertices carry a normal where the
lit formats keep their diffuse, and line vertices are 16; all were strided or
swizzled as garbage. decodeFvf gained XYZRHW support, pipelines are built from
the decoded layout and cached on (fvf, textured, blend), and formats without a
diffuse use new nocolor shader variants rather than reading a normal as colour.

Blend modes now come from the effect table rather than a bool that could only
express straight alpha. Effects 103 and 104 use SRCBLEND=DESTCOLOR with
DESTBLEND=ZERO in CGraphicsEngine::SetShadingEffect -- a multiply -- so the
terrain noise modulates the ground instead of painting flat grey over it.

Pre-transformed positions bypass the world/view/projection product, as D3D9
does for GFXFVF_XYZRHW.

SetTexture no longer forwards stage 1 to the single-sampler renderer, where it
overwrote the stage 0 binding: terrain sets its tileset on 0 and noise on 1, so
every terrain draw sampled the wrong texture.

Mouse motion no longer emits MOUSE_AXIS_X/Y. Those are bound only to the
cursor_x/cursor_y sliders, which CCursor::Update accumulates as relative deltas,
while SDL reports absolute coordinates that the frame already applies with
SetPos. Feeding the absolute position in as a delta moved the cursor twice per
event and made it jitter.

The shader build step took its sources as a plain path argument, so it was
cached on argv alone and an edited shader was never recompiled or staged. Its
sources are declared inputs now.

Still wrong: the renderer binds one sampler, so effects 100/101/102/104 lose
their second stage -- terrain tile transitions draw with only their base texture
and stay hard-edged. GetViewVolume is still a stub, so nothing is culled; a port
of the legacy version culled the whole scene and needs its clip convention
checked. The off-map wedge should be covered by CTerrain::DrawBorder. None of
the shared-code changes are verified on Windows.

+++ Sources/src/Scene/SoundScene.h
@@ -276,7 +276,13 @@ private:
 	{
 		size_t operator() ( const SIntPair &v ) const
 		{
-			return v.x * RAND_MAX +  v.y;
+			// RAND_MAX is 32767 on Windows but INT_MAX here, so the product
+			// overflows a signed int for any x above one and traps. Computing it
+			// unsigned makes the wrap defined and leaves the Windows result
+			// unchanged, since nothing overflowed there. The two platforms still
+			// bucket differently, which matters only if unordered_map iteration
+			// order ever has to agree across them.
+			return static_cast<size_t>( static_cast<unsigned int>( v.x ) * static_cast<unsigned int>( RAND_MAX ) + static_cast<unsigned int>( v.y ) );
 		}
 	};
